refactor: delete root thread history shell
