--- backend/thread_history.py
@@ -1,3 +1,0 @@
-"""Compatibility shell for thread runtime history helpers."""
-
-from backend.thread_runtime.history import *  # noqa: F403
